Rename application sections to steps
Moved all components under application/sections to application/steps and updated imports in apply/page.tsx to reflect the new path. This improves naming consistency for application step components.
